+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added links in resources
 ### Refactor
 - Use DNS bundle product instead of sdk-go-dns
+- Use Container Registry bundle product instead of sdk-go-container-registry
 
 ## 6.7.3
 ### Fixes